oairepository: models addition
* Moves the OaiREPOSITORY model defined under oaiharvester to the
  correct place inside the oairepository module.
